transter loop_over_everything function to outliers.py
